refactor: move models folder one level up
