+++ packages/x-adapter-platform/src/types/models/next-query.model.ts
